Refactor NAIP export notebook
- Remove unused code
- Switch from implicit float32 export to int16 to save space. Ideally
we would use uint8, but that is blocked by an EE bug that prevents
parsing exported byte TFRecords. See https://issuetracker.google.com/issues/296941927.
- Switch from GCP to Google Drive export to avoid ingress/egress costs
while we're developing this. Note that this currently requires a manual
download once exports complete. We may want to automate that process in
the future, but if we were operationalizing this, we would probably be
going back to GCP anyways.