Switch from int16 to uint8 TFRecord export
Parsing uint8 TFRecords is a little tricky as they need to be parsed as
strings first and then decoded to uint8, but it saves a substantial
amount of space.

See https://issuetracker.google.com/issues/296941927 for details.

Note that if we want to support masked data, we'll need another solution
since there's no unused value to use as NODATA.