Predict 2D labels (#11)
The main change here is that we went from predicting single
labels from NAIP arrays to predicting array labels from NAIP
arrays, which should add flexibility if we want to modify output
resolution in the future.

Lots of other changes below:

* Switch LiDAR sampling from local to EE
* Sample all footprints instead of subsampling
* Add Acquisition class for handling LiDAR/NAIP loading
* Update everything to use HDF5 datasets
* Switch from int16 to uint8 TFRecord export
* Add basic UNet model
* Remove ee.Authenticate and autreload from notebooks
* Shuffle datasets before splitting
* Add early stopping and model checkpoint callbacks
* Add and run pre-commit hooks
* Switch from GCP to Google Drive export to avoid ingress/egress costs
while we're developing this. Note that this currently requires a manual
download once exports complete. We may want to automate that process in
the future, but if we were operationalizing this, we would probably be
going back to GCP anyways.