Update dev workflow to use cloned remote repository
Using a local copy of the repository causes syncing issues between
Windows and WSL that prevents file events from being detected. This
ranges from a minor inconvenience (having to manually refresh source
control) to a major issue (DVC experiments not being updated and files
being out of sync).

To fix that, we can avoid using a local copy of the repository, and
instead clone the remote repository during development. Data was
previously stored untracked in the local repository, but it must now be
mounted from a separate location on the local system.