Log model params with W&B
Previously we only logged the model name and had to create a unique
model for each param change. This was unwieldy and required us to
keep all models to avoid losing metadata. Now, models are created
based on params which are logged to Weights and Biases.

In the process, I went back and renamed old runs with new model names
to reflect the fact that most were the same model with different
params. I kept Adam's original model as CNN_v1, while all of my runs
are now permutations of CNN_v2.